+++ emu/processes/wps_poly_centroid.py
@@ -1,9 +1,9 @@
 """
 Process using an application/xml+gml input. Used to test UI interactions.
 
-Author: David Huard
+Author: David Huard and Trevor James Smith
 """
-from pywps import Process, ComplexInput, LiteralOutput
+from pywps import Process, ComplexInput, LiteralInput, LiteralOutput
 from pywps import FORMATS
 import logging
 LOGGER = logging.getLogger("PYWPS")
@@ -12,9 +12,12 @@ LOGGER = logging.getLogger("PYWPS")
 class PolyCentroid(Process):
     def __init__(self):
         inputs = [
-            ComplexInput('polygon', 'Region definition',
+            LiteralInput("wkt", "Region definition in WKT: Well-Known-Text format",
+                         abstract="A Well-Known-Test definition for a region.",
+                         min_occurs=0, data_type="string", default=""),
+            ComplexInput('xml', 'Region definition in XML format',
                          abstract="A polygon defining a region.",
-                         supported_formats=[FORMATS.GML, ]),
+                         min_occurs=0, supported_formats=[FORMATS.GML, ]),
         ]
         outputs = [
             LiteralOutput('output', 'The centroid of the polygon geometry.',
@@ -27,7 +30,7 @@ class PolyCentroid(Process):
             title="Approximate centroid of a polygon.",
             abstract="Return the polygon's centroid coordinates. If the geometry contains multiple polygons, "
                      "only the centroid of the first one will be computed. Do not use for serious computations"
-                     ", this is only a test process and uses a crude approximation. ",
+                     ", this is only a test process and uses a crude approximation.",
             version="1.0",
             inputs=inputs,
             outputs=outputs,
@@ -37,28 +40,55 @@ class PolyCentroid(Process):
 
     @staticmethod
     def _handler(request, response):
+        from geomet import wkt
         from defusedxml import ElementTree
         response.update_status('PyWPS Process started.', 0)
 
-        fn = request.inputs['polygon'][0].file
+        if request.inputs['wkt'][0].data != "":
 
-        ns = {'gml': 'http://www.opengis.net/gml'}
-        poly = ElementTree.parse(fn)
+            try:
+                fn = request.inputs['wkt'][0].data
+                poly = wkt.loads(fn)
 
-        # Find the first polygon in the file.
-        e = poly.find('.//gml:Polygon', ns)
+                # Get the coordinates of the first feature
+                if len(poly['coordinates']) == 1:  # For Polygons and Multipolygons
+                    coordinates = poly['coordinates'][0]
+                else:
+                    coordinates = poly['coordinates']  # For other geometries
 
-        # Get the coordinates
-        c = e.find('.//gml:coordinates', ns).text
-        coords = [tuple(map(float, p.split(','))) for p in c.split(' ')]
+            except Exception as e:
+                msg = "{}: WKT not found.".format(e)
+                logging.warning(msg=msg)
+                raise
+
+        elif request.inputs['xml'][0].file is not None:
+
+            try:
+                fn = request.inputs['xml'][0].file
+                poly = ElementTree.parse(fn)
+                ns = {'gml': 'http://www.opengis.net/gml'}
+
+                # Find the first polygon in the file.
+                e = poly.find('.//gml:Polygon', ns)
+
+                # Get the coordinates
+                c = e.find('.//gml:coordinates', ns).text
+                coordinates = [tuple(map(float, p.split(','))) for p in c.split(' ')]
+
+            except Exception as e:
+                msg = "{}: XML not found.".format(e)
+                logging.warning(msg=msg)
+                raise
+        else:
+            raise ValueError("Process requires a WKT string or XML file.")
 
         # Compute the average
-        n = len(coords)
-        x, y = zip(*coords)
-        cx = sum(x) / n
-        cy = sum(y) / n
+        n = len(coordinates)
+        x, y = zip(*coordinates)
+        centroid_x = sum(x) / n
+        centroid_y = sum(y) / n
 
-        response.outputs['output'].data = '{:.5f},{:.5f}'.format(cx, cy)
+        response.outputs['output'].data = '{:.5f},{:.5f}'.format(centroid_x, centroid_y)
 
         response.update_status('PyWPS Process completed.', 100)
         return response