+++ tests/common.py
@@ -4,6 +4,7 @@ from pywps.app.basic import get_xpath_ns
 from pywps.tests import WpsClient, WpsTestResponse
 
 TESTS_HOME = os.path.abspath(os.path.dirname(__file__))
+CFG_FILE = os.path.join(TESTS_HOME, 'test.cfg')
 
 VERSION = "1.0.0"
 WPS, OWS = get_ElementMakerForVersion(VERSION)
@@ -19,7 +20,7 @@ class WpsTestClient(WpsClient):
     def get(self, *args, **kwargs):
         query = "?"
         for key, value in kwargs.items():
-            query += "{0}={1}&".format(key, value)
+            query += u"{0}={1}&".format(key, value)
         return super(WpsTestClient, self).get(query)